90% done
I have reset all the seeder because i change the database structure which before i make a mistake by making the relation of user to student and institution to 1:N which now is 1:1.

And then I have finished my blockchain to send/create document into the blockchain also to read the verified document

In the view file of documents (admin) when file is uploaded already to the blockchain button will change into view on blockchain. And when the admin click it, it will show the JSON file of the uploaded doc into the blockchain